fix(security): mask sensitive data in clear-text-logging findings (CodeQL HIGH phase-1a)
